% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_state-trace.R
\name{staCMR}
\alias{staCMR}
\alias{continuous_cmr}
\alias{staMR}
\alias{staMRFIT}
\alias{staCMRFIT}
\title{CMR State-Trace Analysis}
\usage{
staCMR(data, partial = list(), shrink = -1, approx = FALSE)

staMR(data = list(), partial = list(), shrink = -1)

staMRFIT(data = NULL, partial = list(), nsample = 1, shrink = -1)

staCMRFIT(data = NULL, partial = list(), nsample = 1, shrink = -1,
  approx = FALSE)
}
\arguments{
\item{data}{Either a data structure (in list or general format, see
\link{gen2list}) or structured output from \link{staSTATS}.}

\item{partial}{is a partial order in either list or adjacency matrix format.}

\item{shrink}{Shrinkage parameter (see \link{staSTATS}). Default calculates
optimum amount of shrinkage.}

\item{approx}{\code{FALSE} (the default) uses full algorithm, \code{TRUE} uses an
approximate algorithm.}

\item{nsample}{no. of Monte Carlo samples (about 10000 is good)}
}
\description{
\code{staCMR} is the main function that conducts the CMR (state-trace) analysis
for continuous data. It takes a data structure or a list of structured output
from \link{staSTATS} and an optional partial order and returns the best fitting
values (to the data means) and the least squares fit. It fits the monotonic
model to the data.

\code{staMR} conducts monotonic regression on a data structure according to a given
partial order. We say it fits the partial order model to the data (i.e., the
set of dependent variables).

\code{staMRFIT} tests the fit of the partial order model.

\code{staCMRFIT}  estimates the empirical distribution (and hence \emph{p}-value) of
the difference in the fit of the conjoint monotonic and the fit of the
partial order model.
}
\examples{

library("STACMR")

## load data from Exp. 1 of Dunn, Newell, & Kalish (2012)
data(delay)
head(delay)

y <- gen2list(data=delay)
str(y, 2)

delaystats <- staSTATS(delay)
str(delaystats)

delaystats[[1]]$means

delaystats[[2]]$means

staPLOT(data = delay, 
        groups = list(c(1:4), c(5:8)), 
        grouplabels = list("No delay", "Delay"), 
        axislabels = list("RB","II"))


### Fits Partial Order Model to Data
# create partial order
E <- list(c(1:4),c(5:8),c(5,1),c(6,2),c(7,3),c(8,4))
out2 <- staMR(data=delay, partial=E)

cbind(out2$x[[1]],out2$x[[2]]) # simplify presentation
out2$fval
out2$shrinkage

### Conduct CMR State-Trace Analysis
out1 <- staCMR(data=delay, partial=E)

cbind(out1$x[[1]],out1$x[[2]])
out1$fval
out1$shrinkage

staPLOT(data = delay, 
        groups = list(c(1:4), c(5:8)), 
        grouplabels = list("No delay", "Delay"), 
        axislabels = list("RB","II"),
        pred=out1$x)

### Test fit of Partial Order Model

out3 <- staMRFIT(delay, partial=E, nsample=10000)
out3$p
out3$datafit

### p-Value of Difference in Fit of Conjoint Monotonic and Fit of Partial Order Model
out4 <- staCMRFIT(delay, partial=E, nsample=10000)
out4$p
out4$datafit

}
