% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen2list.R
\name{gen2list}
\alias{gen2list}
\alias{data_formats}
\title{Data formats}
\usage{
gen2list(data = NULL, varnames)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} in general format (see below).}

\item{varnames}{optional \code{list}  of the names of each within-participant
condition.}
}
\value{
\code{gen2list} returns a \code{ngroup} x \code{nvar} \code{list} in which each element
is an \code{nsub} x \code{ncond} matrix of values
}
\description{
\pkg{STACMR} accepts two kinds of data structure, \strong{list
format} and \strong{general format}. \code{gen2list} transforms from the general
format into the list format.
}
\details{
\subsection{List Format}{In this format, the data are contained in a \code{b} x
\code{n} list where \code{b} is the number of between-participant conditions (groups)
and \code{n} is the number of dependent variables. Each component of the list is
itself an \code{N} x \code{w} matrix of observations where \code{N} is the number of
subjects (which may vary across groups and dependent variables) and \code{w} is
the number of within-participant conditions (fixed across groups and
dependent variables). The dependent variable may be either within-participant
or between-participant. This does not matter because the correlation between
dependent variables is assumed to be zero (although this might change in
future implementations).}

\subsection{General format}{This is a fixed column format organised as a
matrix in which each row corresponds to an observation and each column is
defined as follows:
\enumerate{
\item Participant number (for identification only, not used directly)
\item Between-participant condition or group (if none, then set this value to 1)
\item Dependent variable (numbered 1, 2, and so on)
\item column 4 to end: Values for each within-participant condition
}
}
}
