% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staPLOT.R
\name{staPLOT}
\alias{staPLOT}
\title{State-Trace Plot for Continuous Data}
\usage{
staPLOT(data, vars = c(1, 2), groups = NULL, grouplabels = NULL,
  axislabels = c("DV1", "DV2"), xlim = NULL, ylim = NULL,
  pred = NULL, palette = "Set1")
}
\arguments{
\item{data}{Name of the data set and is the one required argument. It may be
in either general or list format, or in summary statistics form (i.e., the
output of staSTATS).}

\item{groups}{\code{list} identifying the conditions to be distinguished in
the plot with different markers.}

\item{grouplabels}{\code{list} consisting of the labels for the groups
defined by groups to appear in the legend.}

\item{axislabels}{\code{list} that defines the labels on the x-axis and
y-axis, respectively.}

\item{xlim, ylim}{vector of axes limits.}

\item{pred}{\code{list} of predicted values for each condition. Typically,
this is the best-fitting values returned by \code{\link{staMR}} or
\code{\link{staCMR}}.}

\item{palette}{identifies a color palette (see
\code{\link[ggplot2]{scale_colour_brewer}}) for plotting points.}
}
\description{
Generates a state-trace plot for continuous data.
}
