% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staSTATS.R
\name{staSTATS}
\alias{staSTATS}
\title{Statistics for state trace analysis}
\usage{
staSTATS(data, shrink = -1, varnames, warning = FALSE)
}
\arguments{
\item{data}{A \code{list} of \code{lists} (returned from \link{gen2list}) containing
\code{nsub} x \code{ncond} matrices}

\item{shrink}{numeric indicating amount of shrinkage to apply to the
estimated covariance matrix. Generally, the covariance matrix needs to be
shrunk during the bootstrap cycle to avoid ill-conditioning. If \code{shrink = 0} then no shrinkage is applied. If \code{shrink = 1} then maximum shrinkage is
applied. This means that the covariance matrix is diagonalized with all
off-diagonal entries set to zero. If \code{shrink < 0} (the default) then an
optimal shrinkage value is estimated for each within-participant block and
applied according to an algorithm developed by Ledoit and Wolf (2004).}

\item{varnames}{An optional \code{list} of names of within-participant conditions}

\item{if}{warning is \code{TRUE} then a warning is thrown if \code{NA}s are detected.
Default is \code{FALSE}.}
}
\description{
Calculates statistics for state trace analysis.
}
\references{
Ledoit, O. & Wolf, M. (2004). Honey, I shrunk the sample
covariance matrix. \emph{The Journal of Portfolio Management}, 30(4), 110-119.
}
